refactor: extract shared base classes from worker manager adapter implementations
Centralises duplicated heartbeat management, task lifecycle, worker process
setup, and event-loop boilerplate into four base classes:

- BaseHeartbeatManager: shared init/register/on_heartbeat_echo/routine with
  _get_processor_statuses() hook for backend-specific processor reporting
- BaseTaskManager: shared priority queue, semaphore, and task-state machinery
  with abstract _execute_task() and hook points for backend-specific cancel/
  cleanup (_on_task_cancel, _on_task_cleanup)
- BaseWorker: shared multiprocessing.Process lifecycle (run/_run/__initialize,
  signal handling, message dispatch) with abstract factory methods
  _create_heartbeat_manager/_create_task_manager and _get_extra_loops() hook
- BaseWorkerManager: shared async event-loop, heartbeat, command dispatch, and
  signal handling with abstract start_worker/shutdown_workers

Reduces ~1,600 lines of duplicated code across the five implementations.
Shared bugs (e.g. missing null guards in resolve_tasks) are now fixed once.

Closes #722
