+++ src/scaler/worker_manager_adapter/baremetal/native.py
@@ -8,25 +8,18 @@ from typing import Any, Dict, List, Optional, Tuple
 import zmq
 
 from scaler.config.section.native_worker_manager import NativeWorkerManagerConfig, NativeWorkerManagerMode
-from scaler.io import ymq
 from scaler.io.mixins import AsyncConnector
 from scaler.io.utility import create_async_connector, create_async_simple_context
-from scaler.protocol.capnp import (
-    BaseMessage,
-    WorkerManagerCommand,
-    WorkerManagerCommandResponse,
-    WorkerManagerCommandType,
-    WorkerManagerHeartbeat,
-    WorkerManagerHeartbeatEcho,
-)
-from scaler.utility.event_loop import create_async_loop_routine, run_task_forever
+from scaler.protocol.capnp import WorkerManagerCommandResponse
+from scaler.utility.event_loop import run_task_forever
 from scaler.utility.identifiers import WorkerID
 from scaler.worker.worker import Worker
+from scaler.worker_manager_adapter.base_manager import BaseWorkerManager
 
 Status = WorkerManagerCommandResponse.Status
 
 
-class NativeWorkerManager:
+class NativeWorkerManager(BaseWorkerManager):
     def __init__(self, config: NativeWorkerManagerConfig):
         self._address = config.worker_manager_config.scheduler_address
         self._worker_scheduler_address = config.worker_manager_config.effective_worker_scheduler_address
@@ -77,7 +70,7 @@ class NativeWorkerManager:
             socket_type=zmq.DEALER,
             address=self._address,
             bind_or_connect="connect",
-            callback=self.__on_receive_external,
+            callback=self._on_receive_external,
             identity=self._ident,
         )
 
@@ -108,57 +101,25 @@ class NativeWorkerManager:
             worker.start()
             self._workers[worker.identity] = worker
 
-    async def __on_receive_external(self, message: BaseMessage):
-        if isinstance(message, WorkerManagerCommand):
-            await self._handle_command(message)
-
-        elif isinstance(message, WorkerManagerHeartbeatEcho):
-            pass
-
-        else:
-            print(f"Received unknown message type: {type(message)}")
-
-    async def _handle_command(self, command: WorkerManagerCommand):
-        cmd_type = command.command
-        response_status: Status = Status.success
-        worker_ids: List[bytes] = []
-
-        if cmd_type == WorkerManagerCommandType.startWorkers:
-            new_wid, response_status = await self.start_worker()
-            if response_status == Status.success:
-                worker_ids = [bytes(new_wid)]
-        elif cmd_type == WorkerManagerCommandType.shutdownWorkers:
-            response_status = await self.shutdown_workers(command.workerIDs)
-            if response_status == Status.success:
-                worker_ids = list(command.workerIDs)
-        else:
-            raise ValueError("Unknown WorkerManagerCommand")
-
-        await self._connector_external.send(
-            WorkerManagerCommandResponse(
-                command=cmd_type, status=response_status, workerIDs=worker_ids, capabilities=self._capabilities
-            )
-        )
-
-    async def start_worker(self) -> Tuple[WorkerID, Status]:
+    async def start_worker(self) -> Tuple[List[bytes], Status]:
         if len(self._workers) >= self._max_task_concurrency != -1:
-            return WorkerID(b""), Status.tooManyWorkers
+            return [], Status.tooManyWorkers
 
         worker = self._create_worker()
         worker.start()
         self._workers[worker.identity] = worker
         logging.info(f"Start worker, {self._ident!r}")
-        return worker.identity, Status.success
+        return [bytes(worker.identity)], Status.success
 
-    async def shutdown_workers(self, worker_ids: List[bytes]) -> Status:
+    async def shutdown_workers(self, worker_ids: List[bytes]) -> Tuple[List[bytes], Status]:
         if not worker_ids:
-            return Status.workerNotFound
+            return [], Status.workerNotFound
 
         for wid_bytes in worker_ids:
             wid = WorkerID(wid_bytes)
             if wid not in self._workers:
                 logging.warning(f"Worker with ID {wid!r} does not exist.")
-                return Status.workerNotFound
+                return [], Status.workerNotFound
 
         for wid_bytes in worker_ids:
             wid = WorkerID(wid_bytes)
@@ -166,14 +127,13 @@ class NativeWorkerManager:
             os.kill(worker.pid, signal.SIGINT)
             worker.join()
 
-        return Status.success
+        return list(worker_ids), Status.success
 
     def run(self) -> None:
         if self._mode == NativeWorkerManagerMode.FIXED:
             self._run_fixed()
             return
 
-        # DYNAMIC mode
         self._setup_zmq()
         self._loop = asyncio.new_event_loop()
         run_task_forever(self._loop, self._run(), cleanup_callback=self._cleanup)
@@ -192,45 +152,3 @@ class NativeWorkerManager:
 
         for worker in self._workers.values():
             worker.join()
-
-    def _cleanup(self) -> None:
-        if self._connector_external is not None:
-            self._connector_external.destroy()
-
-    def __destroy(self):
-        print(f"Worker manager {self._ident!r} received signal, shutting down")
-        self._task.cancel()
-
-    def __register_signal(self):
-        self._loop.add_signal_handler(signal.SIGINT, self.__destroy)
-        self._loop.add_signal_handler(signal.SIGTERM, self.__destroy)
-
-    async def _run(self) -> None:
-        self._task = self._loop.create_task(self.__get_loops())
-        self.__register_signal()
-        await self._task
-
-    async def __send_heartbeat(self) -> None:
-        await self._connector_external.send(
-            WorkerManagerHeartbeat(
-                maxTaskConcurrency=self._max_task_concurrency,
-                capabilities=self._capabilities,
-                workerManagerID=self._worker_manager_id,
-            )
-        )
-
-    async def __get_loops(self) -> None:
-        loops = [
-            create_async_loop_routine(self._connector_external.routine, 0),
-            create_async_loop_routine(self.__send_heartbeat, self._heartbeat_interval_seconds),
-        ]
-
-        try:
-            await asyncio.gather(*loops)
-        except asyncio.CancelledError:
-            pass
-        except ymq.YMQException as e:
-            if e.code == ymq.ErrorCode.ConnectorSocketClosedByRemoteEnd:
-                pass
-            else:
-                logging.exception(f"{self._ident!r}: failed with unhandled exception:\n{e}")
