+++ src/scaler/worker_manager_adapter/orb_aws_ec2/worker_manager.py
@@ -1,7 +1,6 @@
 import asyncio
 import logging
 import os
-import signal
 import uuid
 from typing import Any, Dict, List, Optional, Tuple
 
@@ -15,20 +14,13 @@ except ModuleNotFoundError as exc:
     raise ModuleNotFoundError('execute "pip install opengris-scaler[orb]" to use ORB AWS EC2 worker Manager') from exc
 
 from scaler.config.section.orb_aws_ec2_worker_adapter import ORBAWSEC2WorkerAdapterConfig
-from scaler.io import ymq
 from scaler.io.mixins import AsyncConnector
 from scaler.io.utility import create_async_connector, create_async_simple_context
-from scaler.protocol.capnp import (
-    BaseMessage,
-    WorkerManagerCommand,
-    WorkerManagerCommandResponse,
-    WorkerManagerCommandType,
-    WorkerManagerHeartbeat,
-    WorkerManagerHeartbeatEcho,
-)
-from scaler.utility.event_loop import create_async_loop_routine, register_event_loop, run_task_forever
+from scaler.protocol.capnp import WorkerManagerCommandResponse
+from scaler.utility.event_loop import register_event_loop
 from scaler.utility.identifiers import WorkerID
 from scaler.utility.logging.utility import setup_logger
+from scaler.worker_manager_adapter.base_manager import BaseWorkerManager
 from scaler.worker_manager_adapter.common import format_capabilities
 
 Status = WorkerManagerCommandResponse.Status
@@ -50,7 +42,7 @@ def get_orb_aws_ec2_worker_name(instance_id: str) -> str:
     return f"Worker|ORB|{instance_id}|{tag}"
 
 
-class ORBAWSEC2WorkerAdapter:
+class ORBAWSEC2WorkerAdapter(BaseWorkerManager):
     _config: ORBAWSEC2WorkerAdapterConfig
     _sdk: Optional[Any]
     _workers: Dict[WorkerID, str]
@@ -76,7 +68,6 @@ class ORBAWSEC2WorkerAdapter:
 
         self._sdk: Optional[Any] = None
         self._ec2: Optional[Any] = None
-        self._context = None
         self._connector_external: Optional[AsyncConnector] = None
         self._created_security_group_id: Optional[str] = None
         self._created_key_name: Optional[str] = None
@@ -115,7 +106,7 @@ class ORBAWSEC2WorkerAdapter:
             "storage": {"type": "json"},
         }
 
-    async def __setup(self) -> None:
+    async def _setup(self) -> None:
         """Set up AWS resources and the ORB template after the SDK is initialised."""
         region = self._config.aws_region or "us-east-1"
         self._ec2 = boto3.client("ec2", region_name=region)
@@ -155,21 +146,21 @@ class ORBAWSEC2WorkerAdapter:
         validate_result = await self._sdk.validate_template(template_id=self._template_id)
         logging.info(f"validate_template result: {validate_result}")
 
-        self._context = create_async_simple_context()
+        context = create_async_simple_context()
         self._name = "worker_manager_orb_aws_ec2"
         self._ident = f"{self._name}|{uuid.uuid4().bytes.hex()}".encode()
 
         self._connector_external = create_async_connector(
-            self._context,
+            context,
             name=self._name,
             socket_type=zmq.DEALER,
             address=self._address,
             bind_or_connect="connect",
-            callback=self.__on_receive_external,
+            callback=self._on_receive_external,
             identity=self._ident,
         )
 
-    async def __terminate_all_workers(self) -> None:
+    async def _terminate_all_workers(self) -> None:
         """Return all active instances to ORB before the SDK context exits."""
         if not self._workers or self._sdk is None:
             return
@@ -182,61 +173,11 @@ class ORBAWSEC2WorkerAdapter:
             logging.warning(f"Failed to terminate instances during cleanup: {e}")
         self._workers.clear()
 
-    async def __on_receive_external(self, message: BaseMessage):
+    async def _on_receive_external(self, message) -> None:
         try:
-            if isinstance(message, WorkerManagerCommand):
-                await self._handle_command(message)
-            elif isinstance(message, WorkerManagerHeartbeatEcho):
-                pass
-            else:
-                logging.warning(f"Received unknown message type: {type(message)}")
+            await super()._on_receive_external(message)
         except Exception:
             logging.exception(f"Unhandled exception while processing message {type(message).__name__}")
-
-    async def _handle_command(self, command: WorkerManagerCommand):
-        cmd_type = command.command
-        response_status: Status = Status.success
-        worker_ids: List[bytes] = []
-        capabilities: Dict[str, int] = {}
-
-        if cmd_type == WorkerManagerCommandType.startWorkers:
-            worker_ids, response_status = await self.start_worker()
-            if response_status == Status.success:
-                capabilities = self._capabilities
-        elif cmd_type == WorkerManagerCommandType.shutdownWorkers:
-            worker_ids, response_status = await self.shutdown_workers(list(command.workerIDs))
-        else:
-            logging.error(f"Received unknown command type: {cmd_type!r}")
-            raise ValueError(f"Unknown Command: {cmd_type!r}")
-
-        assert self._connector_external is not None
-        await self._connector_external.send(
-            WorkerManagerCommandResponse(
-                command=cmd_type, status=response_status, workerIDs=worker_ids, capabilities=capabilities
-            )
-        )
-
-    async def __send_heartbeat(self):
-        assert self._connector_external is not None
-        await self._connector_external.send(
-            WorkerManagerHeartbeat(
-                maxTaskConcurrency=self._max_task_concurrency,
-                capabilities=self._capabilities,
-                workerManagerID=self._worker_manager_id,
-            )
-        )
-
-    def run(self) -> None:
-        self._loop = asyncio.new_event_loop()
-        run_task_forever(self._loop, self._run(), cleanup_callback=self._cleanup)
-
-    def __destroy(self):
-        logging.info(f"Worker adapter {self._ident!r} received signal, shutting down")
-        self._task.cancel()
-
-    def __register_signal(self):
-        self._loop.add_signal_handler(signal.SIGINT, self.__destroy)
-        self._loop.add_signal_handler(signal.SIGTERM, self.__destroy)
 
     async def _run(self) -> None:
         register_event_loop(self._event_loop)
@@ -253,190 +194,19 @@ class ORBAWSEC2WorkerAdapter:
             # setup_logger is called after the ORB context is entered because ORB reconfigures
             # the root logger during __aenter__, which would otherwise suppress scaler log output.
             setup_logger(self._logging_paths, self._logging_config_file, self._logging_level)
-            await self.__setup()
-            self._task = self._loop.create_task(self.__get_loops())
-            self.__register_signal()
+            await self._setup()
+            self._task = self._loop.create_task(self._get_loops())
+            self._register_signal()
             try:
                 await self._task
             except asyncio.CancelledError:
                 pass
             finally:
-                await self.__terminate_all_workers()
+                await self._terminate_all_workers()
 
         self._sdk = None
 
-    async def __get_loops(self):
-        assert self._connector_external is not None
-        loops = [
-            create_async_loop_routine(self._connector_external.routine, 0),
-            create_async_loop_routine(self.__send_heartbeat, self._heartbeat_interval_seconds),
-        ]
-
-        try:
-            await asyncio.gather(*loops)
-        except asyncio.CancelledError:
-            pass
-        except ymq.YMQException as e:
-            if e.code == ymq.ErrorCode.ConnectorSocketClosedByRemoteEnd:
-                pass
-            else:
-                logging.exception(f"{self._ident!r}: failed with unhandled exception:\n{e}")
-        except Exception:
-            logging.exception(f"{self._ident!r}: failed with unhandled exception")
-
-    @staticmethod
-    def _load_requirements_content(requirements_txt: str) -> str:
-        """Return requirements content from a file path or a literal string."""
-        if os.path.isfile(requirements_txt):
-            with open(requirements_txt) as f:
-                return f.read()
-        return requirements_txt
-
-    @staticmethod
-    def _validate_requirements(requirements_content: str) -> None:
-        """Raise ValueError if the requirements content is invalid or does not include opengris-scaler.
-
-        Each non-comment, non-flag line must be parseable as a PEP 508 requirement or a direct URL
-        reference (containing '://'). Lines that fail both checks would cause `pip install` to fail
-        inside the EC2 userdata script.
-        """
-        found_scaler = False
-        for line in requirements_content.splitlines():
-            line = line.strip()
-            if not line or line.startswith("#") or line.startswith("-"):
-                continue
-            try:
-                req = Requirement(line)
-                if canonicalize_name(req.name) == "opengris-scaler":
-                    found_scaler = True
-            except Exception:
-                if "://" not in line:
-                    raise ValueError(f"Invalid requirement line that would cause pip to fail: {line!r}")
-
-        if not found_scaler:
-            raise ValueError(
-                "The requirements file must include the 'opengris-scaler' package. "
-                "Workers will fail to start without it."
-            )
-
-    def _create_user_data(self) -> str:
-        worker_config = self._config.worker_config
-        adapter_config = self._config.worker_manager_config
-
-        script = "#!/bin/bash\n"
-
-        if self._config.image_id is None:
-            python_version = self._config.python_version
-            requirements_txt = self._config.requirements_txt
-
-            requirements_content = self._load_requirements_content(requirements_txt)
-
-            # Phase 1: install Python and dependencies. User data runs as root so no sudo is needed.
-            # set -e ensures any install failure aborts the script rather than launching a broken worker.
-            script += f"""set -e
-dnf update -y
-dnf install -y python{python_version} python{python_version}-pip
-python{python_version} -m venv /opt/opengris-scaler
-/opt/opengris-scaler/bin/python -m pip install --upgrade pip
-cat > /tmp/requirements.txt << 'REQUIREMENTS_EOF'
-{requirements_content}
-REQUIREMENTS_EOF
-/opt/opengris-scaler/bin/pip install -r /tmp/requirements.txt
-ln -sf /opt/opengris-scaler/bin/scaler_* /usr/local/bin/
-set +e
-
-"""
-
-        # Phase 2: launch the worker manager.
-        # NOTE: --max-task-concurrency is not passed; scaler_worker_manager defaults to cpu_count - 1 workers,
-        # where cpu_count is determined by the machine type configured by the user.
-        script += f"""INSTANCE_ID=$(ec2-metadata --instance-id --quiet)
-nohup scaler_worker_manager baremetal_native {self._worker_scheduler_address.to_address()} \
-    --mode fixed \
-    --worker-type ORB \
-    --worker-manager-id "${{INSTANCE_ID}}" \
-    --per-worker-task-queue-size {worker_config.per_worker_task_queue_size} \
-    --heartbeat-interval-seconds {worker_config.heartbeat_interval_seconds} \
-    --task-timeout-seconds {worker_config.task_timeout_seconds} \
-    --garbage-collect-interval-seconds {worker_config.garbage_collect_interval_seconds} \
-    --death-timeout-seconds {worker_config.death_timeout_seconds} \
-    --trim-memory-threshold-bytes {worker_config.trim_memory_threshold_bytes} \
-    --event-loop {self._config.worker_config.event_loop} \
-    --io-threads {self._config.worker_config.io_threads}"""
-
-        if worker_config.hard_processor_suspend:
-            script += " \
-    --hard-processor-suspend"
-
-        if adapter_config.object_storage_address:
-            script += f" \
-    --object-storage-address {adapter_config.object_storage_address.to_string()}"
-
-        capabilities = worker_config.per_worker_capabilities.capabilities
-        if capabilities:
-            cap_str = format_capabilities(capabilities)
-            if cap_str.strip():
-                script += f" \
-    --per-worker-capabilities {cap_str}"
-
-        script += " > /var/log/opengris-scaler.log 2>&1 &\n"
-
-        return script
-
-    def _discover_latest_al2023_ami(self) -> str:
-        """Discover the most recent Amazon Linux 2023 x86_64 EBS HVM AMI in the configured region."""
-        response = self._ec2.describe_images(
-            Filters=[
-                {"Name": "name", "Values": ["al2023-ami-2023.*-kernel-*-x86_64"]},
-                {"Name": "root-device-type", "Values": ["ebs"]},
-                {"Name": "virtualization-type", "Values": ["hvm"]},
-            ],
-            Owners=["amazon"],
-        )
-        images = response.get("Images", [])
-        if not images:
-            raise RuntimeError("No AL2023 AMI found in the current region.")
-        images.sort(key=lambda img: img["CreationDate"], reverse=True)
-        ami_id = images[0]["ImageId"]
-        logging.info(f"Auto-discovered latest AL2023 AMI: {ami_id}")
-        return ami_id
-
-    def _discover_default_subnet(self) -> str:
-        vpcs = self._ec2.describe_vpcs(Filters=[{"Name": "isDefault", "Values": ["true"]}])
-        if not vpcs["Vpcs"]:
-            raise RuntimeError("No default VPC found, and no subnet_id provided.")
-        default_vpc_id = vpcs["Vpcs"][0]["VpcId"]
-
-        subnets = self._ec2.describe_subnets(Filters=[{"Name": "vpc-id", "Values": [default_vpc_id]}])
-        if not subnets["Subnets"]:
-            raise RuntimeError(f"No subnets found in default VPC {default_vpc_id}.")
-
-        subnet_id = subnets["Subnets"][0]["SubnetId"]
-        logging.info(f"Auto-discovered subnet_id: {subnet_id}")
-        return subnet_id
-
-    def _create_security_group(self):
-        # Get VPC ID from Subnet
-        subnet_response = self._ec2.describe_subnets(SubnetIds=[self._subnet_id])
-        vpc_id = subnet_response["Subnets"][0]["VpcId"]
-
-        # Create Security Group (outbound-only — workers connect out to scheduler via ZMQ)
-        group_name = f"opengris-orb-sg-{self._template_id}"
-        sg_response = self._ec2.create_security_group(
-            Description="Temporary security group created for OpenGRIS ORB worker adapter",
-            GroupName=group_name,
-            VpcId=vpc_id,
-        )
-        self._created_security_group_id = sg_response["GroupId"]
-        logging.info(f"Created security group with ID: {self._created_security_group_id}")
-
-    def _create_key_pair(self):
-        key_name = f"opengris-orb-key-{self._template_id}"
-        self._ec2.create_key_pair(KeyName=key_name)
-        self._created_key_name = key_name
-        logging.info(f"Created key pair: {key_name}")
-
-    def _cleanup(self):
+    def _cleanup(self) -> None:
         if self._cleaned_up:
             return
         self._cleaned_up = True
@@ -462,7 +232,7 @@ nohup scaler_worker_manager baremetal_native {self._worker_scheduler_address.to_
 
         logging.info("Cleanup completed.")
 
-    def __del__(self):
+    def __del__(self) -> None:
         self._cleanup()
 
     async def start_worker(self) -> Tuple[List[bytes], Status]:
@@ -549,3 +319,145 @@ nohup scaler_worker_manager baremetal_native {self._worker_scheduler_address.to_
 
         logging.info(f"Successfully stopped {len(affected_worker_ids)} worker(s): instances {instance_ids}")
         return affected_worker_ids, Status.success
+
+    @staticmethod
+    def _load_requirements_content(requirements_txt: str) -> str:
+        """Return requirements content from a file path or a literal string."""
+        if os.path.isfile(requirements_txt):
+            with open(requirements_txt) as f:
+                return f.read()
+        return requirements_txt
+
+    @staticmethod
+    def _validate_requirements(requirements_content: str) -> None:
+        """Raise ValueError if the requirements content is invalid or does not include opengris-scaler.
+
+        Each non-comment, non-flag line must be parseable as a PEP 508 requirement or a direct URL
+        reference (containing '://'). Lines that fail both checks would cause `pip install` to fail
+        inside the EC2 userdata script.
+        """
+        found_scaler = False
+        for line in requirements_content.splitlines():
+            line = line.strip()
+            if not line or line.startswith("#") or line.startswith("-"):
+                continue
+            try:
+                req = Requirement(line)
+                if canonicalize_name(req.name) == "opengris-scaler":
+                    found_scaler = True
+            except Exception:
+                if "://" not in line:
+                    raise ValueError(f"Invalid requirement line that would cause pip to fail: {line!r}")
+
+        if not found_scaler:
+            raise ValueError(
+                "The requirements file must include the 'opengris-scaler' package. "
+                "Workers will fail to start without it."
+            )
+
+    def _create_user_data(self) -> str:
+        worker_config = self._config.worker_config
+        adapter_config = self._config.worker_manager_config
+
+        script = "#!/bin/bash\n"
+
+        if self._config.image_id is None:
+            python_version = self._config.python_version
+            requirements_txt = self._config.requirements_txt
+
+            requirements_content = self._load_requirements_content(requirements_txt)
+
+            script += f"""set -e
+dnf update -y
+dnf install -y python{python_version} python{python_version}-pip
+python{python_version} -m venv /opt/opengris-scaler
+/opt/opengris-scaler/bin/python -m pip install --upgrade pip
+cat > /tmp/requirements.txt << 'REQUIREMENTS_EOF'
+{requirements_content}
+REQUIREMENTS_EOF
+/opt/opengris-scaler/bin/pip install -r /tmp/requirements.txt
+ln -sf /opt/opengris-scaler/bin/scaler_* /usr/local/bin/
+set +e
+
+"""
+
+        script += f"""INSTANCE_ID=$(ec2-metadata --instance-id --quiet)
+nohup scaler_worker_manager baremetal_native {self._worker_scheduler_address.to_address()} \\
+    --mode fixed \\
+    --worker-type ORB \\
+    --worker-manager-id "${{INSTANCE_ID}}" \\
+    --per-worker-task-queue-size {worker_config.per_worker_task_queue_size} \\
+    --heartbeat-interval-seconds {worker_config.heartbeat_interval_seconds} \\
+    --task-timeout-seconds {worker_config.task_timeout_seconds} \\
+    --garbage-collect-interval-seconds {worker_config.garbage_collect_interval_seconds} \\
+    --death-timeout-seconds {worker_config.death_timeout_seconds} \\
+    --trim-memory-threshold-bytes {worker_config.trim_memory_threshold_bytes} \\
+    --event-loop {self._config.worker_config.event_loop} \\
+    --io-threads {self._config.worker_config.io_threads}"""
+
+        if worker_config.hard_processor_suspend:
+            script += " \\\n    --hard-processor-suspend"
+
+        if adapter_config.object_storage_address:
+            script += f" \\\n    --object-storage-address {adapter_config.object_storage_address.to_string()}"
+
+        capabilities = worker_config.per_worker_capabilities.capabilities
+        if capabilities:
+            cap_str = format_capabilities(capabilities)
+            if cap_str.strip():
+                script += f" \\\n    --per-worker-capabilities {cap_str}"
+
+        script += " > /var/log/opengris-scaler.log 2>&1 &\n"
+
+        return script
+
+    def _discover_latest_al2023_ami(self) -> str:
+        """Discover the most recent Amazon Linux 2023 x86_64 EBS HVM AMI in the configured region."""
+        response = self._ec2.describe_images(
+            Filters=[
+                {"Name": "name", "Values": ["al2023-ami-2023.*-kernel-*-x86_64"]},
+                {"Name": "root-device-type", "Values": ["ebs"]},
+                {"Name": "virtualization-type", "Values": ["hvm"]},
+            ],
+            Owners=["amazon"],
+        )
+        images = response.get("Images", [])
+        if not images:
+            raise RuntimeError("No AL2023 AMI found in the current region.")
+        images.sort(key=lambda img: img["CreationDate"], reverse=True)
+        ami_id = images[0]["ImageId"]
+        logging.info(f"Auto-discovered latest AL2023 AMI: {ami_id}")
+        return ami_id
+
+    def _discover_default_subnet(self) -> str:
+        vpcs = self._ec2.describe_vpcs(Filters=[{"Name": "isDefault", "Values": ["true"]}])
+        if not vpcs["Vpcs"]:
+            raise RuntimeError("No default VPC found, and no subnet_id provided.")
+        default_vpc_id = vpcs["Vpcs"][0]["VpcId"]
+
+        subnets = self._ec2.describe_subnets(Filters=[{"Name": "vpc-id", "Values": [default_vpc_id]}])
+        if not subnets["Subnets"]:
+            raise RuntimeError(f"No subnets found in default VPC {default_vpc_id}.")
+
+        subnet_id = subnets["Subnets"][0]["SubnetId"]
+        logging.info(f"Auto-discovered subnet_id: {subnet_id}")
+        return subnet_id
+
+    def _create_security_group(self) -> None:
+        subnet_response = self._ec2.describe_subnets(SubnetIds=[self._subnet_id])
+        vpc_id = subnet_response["Subnets"][0]["VpcId"]
+
+        group_name = f"opengris-orb-sg-{self._template_id}"
+        sg_response = self._ec2.create_security_group(
+            Description="Temporary security group created for OpenGRIS ORB worker adapter",
+            GroupName=group_name,
+            VpcId=vpc_id,
+        )
+        self._created_security_group_id = sg_response["GroupId"]
+        logging.info(f"Created security group with ID: {self._created_security_group_id}")
+
+    def _create_key_pair(self) -> None:
+        key_name = f"opengris-orb-key-{self._template_id}"
+        self._ec2.create_key_pair(KeyName=key_name)
+        self._created_key_name = key_name
+        logging.info(f"Created key pair: {key_name}")
