+++ src/scaler/worker_manager_adapter/symphony/task_manager.py
@@ -1,30 +1,11 @@
 import asyncio
 import logging
 from concurrent.futures import Future
-from typing import Dict, Optional, Set, cast
 
 import cloudpickle
-from bidict import bidict
 
-from scaler import Serializer
-from scaler.io.mixins import AsyncConnector, AsyncObjectStorageConnector
-from scaler.protocol.capnp import (
-    ObjectInstruction,
-    ObjectMetadata,
-    ObjectStorageAddress,
-    Task,
-    TaskCancel,
-    TaskCancelConfirm,
-    TaskCancelConfirmType,
-    TaskResult,
-    TaskResultType,
-)
-from scaler.utility.identifiers import ObjectID, TaskID
-from scaler.utility.metadata.task_flags import retrieve_task_flags_from_task
-from scaler.utility.mixins import Looper
-from scaler.utility.queues.async_priority_queue import AsyncPriorityQueue
-from scaler.utility.serialization import serialize_failure
-from scaler.worker.agent.mixins import HeartbeatManager, TaskManager
+from scaler.protocol.capnp import Task
+from scaler.worker_manager_adapter.base_task_manager import BaseTaskManager
 from scaler.worker_manager_adapter.symphony.callback import SessionCallback
 from scaler.worker_manager_adapter.symphony.message import SoamMessage
 
@@ -34,36 +15,12 @@ except ImportError:
     raise ImportError("IBM Spectrum Symphony API not found, please install it with 'pip install soamapi'.")
 
 
-class SymphonyTaskManager(Looper, TaskManager):
+class SymphonyTaskManager(BaseTaskManager):
     def __init__(self, base_concurrency: int, service_name: str):
-        if isinstance(base_concurrency, int) and base_concurrency <= 0:
-            raise ValueError(f"base_concurrency must be a possible integer, got {base_concurrency}")
+        super().__init__(base_concurrency=base_concurrency)
 
-        self._base_concurrency = base_concurrency
         self._service_name = service_name
 
-        self._executor_semaphore = asyncio.Semaphore(value=self._base_concurrency)
-
-        self._task_id_to_task: Dict[TaskID, Task] = dict()
-        self._task_id_to_future: bidict[TaskID, asyncio.Future] = bidict()
-
-        self._serializers: Dict[bytes, Serializer] = dict()
-
-        self._queued_task_id_queue = AsyncPriorityQueue()
-        self._queued_task_ids: Set[bytes] = set()
-
-        self._acquiring_task_ids: Set[TaskID] = set()  # tasks contesting the semaphore
-        self._processing_task_ids: Set[TaskID] = set()
-        self._canceled_task_ids: Set[TaskID] = set()
-
-        self._object_storage_address: Optional[ObjectStorageAddress] = None
-
-        self._connector_external: Optional[AsyncConnector] = None
-        self._connector_storage: Optional[AsyncObjectStorageConnector] = None
-
-        """
-        SOAM specific code
-        """
         soamapi.initialize()
 
         self._session_callback = SessionCallback()
@@ -81,191 +38,12 @@ class SymphonyTaskManager(Looper, TaskManager):
         self._ibm_soam_session = self._ibm_soam_connection.create_session(ibm_soam_session_attr)
         logging.info(f"established IBM Spectrum Symphony session {self._ibm_soam_session.get_id()}")
 
-    def register(
-        self,
-        connector_external: AsyncConnector,
-        connector_storage: AsyncObjectStorageConnector,
-        heartbeat_manager: HeartbeatManager,
-    ):
-        self._connector_external = connector_external
-        self._connector_storage = connector_storage
-        self._heartbeat_manager = heartbeat_manager
-
-    async def routine(self):  # SymphonyTaskManager has two loops
+    async def routine(self) -> None:
         pass
 
-    async def on_object_instruction(self, instruction: ObjectInstruction):
-        if instruction.instructionType == ObjectInstruction.ObjectInstructionType.delete:
-            for object_id in instruction.objectMetadata.objectIds:
-                self._serializers.pop(object_id, None)  # we only cache serializers
+    async def _execute_task(self, task: Task) -> asyncio.Future:
+        function, arg_objects = await self._load_task_inputs(task)
 
-            return
-
-        logging.error(f"worker received unknown object instruction type {instruction=}")
-
-    async def on_task_new(self, task: Task):
-        task_priority = self.__get_task_priority(task)
-
-        # if semaphore is locked, check if task is higher priority than all acquired tasks
-        # if so, bypass acquiring and execute the task immediately
-        if self._executor_semaphore.locked():
-            for acquired_task_id in self._acquiring_task_ids:
-                acquired_task = self._task_id_to_task[acquired_task_id]
-                acquired_task_priority = self.__get_task_priority(acquired_task)
-                if task_priority <= acquired_task_priority:
-                    break
-            else:
-                self._task_id_to_task[task.taskId] = task
-                self._processing_task_ids.add(task.taskId)
-                self._task_id_to_future[task.taskId] = await self.__execute_task(task)
-                return
-
-        self._task_id_to_task[task.taskId] = task
-        self._queued_task_id_queue.put_nowait((-task_priority, task.taskId))
-        self._queued_task_ids.add(task.taskId)
-
-    async def on_cancel_task(self, task_cancel: TaskCancel):
-        task_queued = task_cancel.taskId in self._queued_task_ids
-        task_processing = task_cancel.taskId in self._processing_task_ids
-
-        if not task_queued and not task_processing:
-            await self._connector_external.send(
-                TaskCancelConfirm(taskId=task_cancel.taskId, cancelConfirmType=TaskCancelConfirmType.cancelNotFound)
-            )
-            return
-
-        if task_processing and not task_cancel.flags.force:
-            await self._connector_external.send(
-                TaskCancelConfirm(taskId=task_cancel.taskId, cancelConfirmType=TaskCancelConfirmType.cancelFailed)
-            )
-            return
-
-        if task_queued:
-            self._queued_task_ids.remove(task_cancel.taskId)
-            self._queued_task_id_queue.remove(task_cancel.taskId)
-
-            # task can be discarded because task was never submitted
-            self._task_id_to_task.pop(task_cancel.taskId)
-
-        if task_processing:
-            future = self._task_id_to_future[task_cancel.taskId]
-            future.cancel()
-
-            # regardless of the future being canceled, the task is considered canceled and cleanup will occur later
-            self._processing_task_ids.remove(task_cancel.taskId)
-            self._canceled_task_ids.add(task_cancel.taskId)
-
-        result = TaskCancelConfirm(taskId=task_cancel.taskId, cancelConfirmType=TaskCancelConfirmType.canceled)
-        await self._connector_external.send(result)
-
-    async def on_task_result(self, result: TaskResult):
-        if result.taskId in self._queued_task_ids:
-            self._queued_task_ids.remove(result.taskId)
-            self._queued_task_id_queue.remove(result.taskId)
-
-        self._processing_task_ids.remove(result.taskId)
-        self._task_id_to_task.pop(result.taskId)
-
-        await self._connector_external.send(result)
-
-    def get_queued_size(self):
-        return self._queued_task_id_queue.qsize()
-
-    def can_accept_task(self):
-        return not self._executor_semaphore.locked()
-
-    async def resolve_tasks(self):
-        if not self._task_id_to_future:
-            await asyncio.sleep(0)
-            return
-
-        done, _ = await asyncio.wait(self._task_id_to_future.values(), return_when=asyncio.FIRST_COMPLETED)
-        for future in done:
-            task_id = self._task_id_to_future.inv.pop(future)
-            task = self._task_id_to_task[task_id]
-
-            if task_id in self._processing_task_ids:
-                self._processing_task_ids.remove(task_id)
-
-                if future.exception() is None:
-                    serializer_id = ObjectID.generate_serializer_object_id(task.source)
-                    serializer = self._serializers[serializer_id]
-                    result_bytes = serializer.serialize(future.result())
-                    result_type = TaskResultType.success
-                else:
-                    result_bytes = serialize_failure(cast(Exception, future.exception()))
-                    result_type = TaskResultType.failed
-
-                result_object_id = ObjectID.generate_object_id(task.source)
-
-                await self._connector_storage.set_object(result_object_id, result_bytes)
-                await self._connector_external.send(
-                    ObjectInstruction(
-                        instructionType=ObjectInstruction.ObjectInstructionType.create,
-                        objectUser=task.source,
-                        objectMetadata=ObjectMetadata(
-                            objectIds=(result_object_id,),
-                            objectTypes=(ObjectMetadata.ObjectContentType.object,),
-                            objectNames=(f"<res {result_object_id.hex()[:6]}>".encode(),),
-                        ),
-                    )
-                )
-
-                await self._connector_external.send(
-                    TaskResult(taskId=task_id, resultType=result_type, metadata=b"", results=[bytes(result_object_id)])
-                )
-
-            elif task_id in self._canceled_task_ids:
-                self._canceled_task_ids.remove(task_id)
-
-            else:
-                raise ValueError(f"task_id {task_id.hex()} not found in processing or canceled tasks")
-
-            if task_id in self._acquiring_task_ids:
-                self._acquiring_task_ids.remove(task_id)
-                self._executor_semaphore.release()
-
-            self._task_id_to_task.pop(task_id)
-
-    async def process_task(self):
-        await self._executor_semaphore.acquire()
-
-        _, task_id = await self._queued_task_id_queue.get()
-        task = self._task_id_to_task[task_id]
-
-        self._acquiring_task_ids.add(task_id)
-        self._processing_task_ids.add(task_id)
-        self._task_id_to_future[task.taskId] = await self.__execute_task(task)
-
-    async def __execute_task(self, task: Task) -> asyncio.Future:
-        """
-        This method is not very efficient because it does let objects linger in the cache. Each time inputs are
-        requested, all object data are requested.
-        """
-        serializer_id = ObjectID.generate_serializer_object_id(task.source)
-
-        if serializer_id not in self._serializers:
-            serializer_bytes = await self._connector_storage.get_object(serializer_id)
-            serializer = cloudpickle.loads(serializer_bytes)
-            self._serializers[serializer_id] = serializer
-        else:
-            serializer = self._serializers[serializer_id]
-
-        # Fetches the function object and the argument objects concurrently
-
-        get_tasks = [
-            self._connector_storage.get_object(object_id)
-            for object_id in [ObjectID(task.funcObjectId), *(ObjectID(argument.data) for argument in task.functionArgs)]
-        ]
-
-        function_bytes, *arg_bytes = await asyncio.gather(*get_tasks)
-
-        function = serializer.deserialize(function_bytes)
-        arg_objects = [serializer.deserialize(object_bytes) for object_bytes in arg_bytes]
-
-        """
-        SOAM specific code
-        """
         input_message = SoamMessage()
         input_message.set_payload(cloudpickle.dumps((function, *arg_objects)))
 
@@ -281,12 +59,3 @@ class SymphonyTaskManager(Looper, TaskManager):
             self._session_callback.submit_task(symphony_task.get_id(), future)
 
         return asyncio.wrap_future(future)
-
-    @staticmethod
-    def __get_task_priority(task: Task) -> int:
-        priority = retrieve_task_flags_from_task(task).priority
-
-        if priority < 0:
-            raise ValueError(f"invalid task priority, must be positive or zero, got {priority}")
-
-        return priority
